# openapi.yaml
openapi: 3.1.0
info:
    title: MLB Stats
    description: Get current and historical statistics for MLB players, teams, games, and seasons.
    version: "v1"
servers:
    - url: https://baseball-stats.azurewebsites.net/
paths:
    /standings:
        get:
            operationId: getStandings
            summary: Retrieves the overall standings for all teams across the MLB depending on the year that is supplied.
            description: The operation will retrieve the ranking of every team within their respective division across Major League Baseball (MLB). 
            parameters:
                - in: query
                  name: year
                  required: true
                  schema:
                    type: integer
                  description: Used to filter the year the overall standings should be returned for.
            responses:
                "200":
                    description: Returns the overall league standings.
                    content:
                      application/json:
                        schema: 
                            type: array
    /news:
        get:
            operationId: getNews
            summary: Retrieves up to date news articles for teams from mlb.com
            parameters:
              - in: query
                name: team_name
                required: true
                schema:
                  type: string
                  enum: ['whitesox', 'guardians', 'tigers', 'royals', 'twins' , 'cubs' , 'reds' , 'brewers' , 'pirates', 
                        'cardinals', 'orioles', 'redsox', 'yankees', 'rays', 'bluejays', 'braves', 'marlins', 'mets', 
                        'phillies', 'nationals', 'astros', 'angels', 'athletics', 'mariners', 'rangers', 'diamondbacks', 
                        'rockies', 'dodgers', 'padres', 'giants']
                description: Used to filter out relevant news articles for the team being specified.
            responses:
              "200":
                description: OK
                content:
                  application/json:
                    schemas:
                      type: array
    /batting_stats_individual:
      get:
          operationId: getBattingStatsIndividual
          summary: Retrieves batting statistics for an individual player for a given season
          parameters:
            - in: query
              name: year
              required: true
              schema:
                type: integer
              description: Used to filter the player's batting statistics based on the year specified.
            - in: query
              name: key_fangraphs
              required: true
              schema:
                type: integer
              description: >
                  The fangraphs player identifier. This value can be retrieved by invoking 
                  the operation with operationId 'getPlayerIdLookup'.
          responses:
            "200":
              description: OK
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/getBattingStatsIndividualResponse'
    /batting_stats:
      get:
          operationId: getBattingStats
          summary: Retrieves a batting statistic for all players from Fangraphs for the season being inputted
          parameters:
            - in: query
              name: year
              required: true
              schema:
                type: integer
              description: Used to filter batting statistics based on the year being specified.
            - in: query
              name: batting_stat
              required: true
              schema:
                type: string
                enum: ['H', '2B', '3B', 'HR', 'RBI', 'BB', 'IBB', 'SO', 'HBP', 'SH', 'SF', 'GDP', 'SB', 'CS', 'AVG', 
                      'OBP', 'SLG', 'OPS']
              description: Used to filter batting statistics based on the statistic being specified.
          responses:
            "200":
              description: >
                Successful operation. The response is a sorted JSON object with the property "Name" 
                and another property which is the value of the input "batting_stat". 
              content:
                application/json:
                  schema:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: The player's name
                        batting_stat: 
                          type: number
                          description: The value of the batting stat inserted as an input to the function.
    /starting_pitching_stats:
      get:
          operationId: getPitchingStatsFangraphs
          summary: Retrieves a specific pitching statistic for all starting pitchers from Fangraphs for the season being inputted.
          parameters:
            - in: query
              name: year
              required: true
              schema: 
                type: integer
              description: Used to select which season the pitching statistics will be retrieved from. 
            - in: query
              name: pitching_stat 
              required: true
              schema:
                type: string
                enum: ['W', 'L', 'WAR', 'ERA', 'GS', 'IP', 'HR', 'BB' , 'SO', 'GB', 'FB', 'LD', 'K/9', 'BB/9', 'H/9', 'AVG',
                      'WHIP', 'BABIP', 'LOB%', 'FIP']
              description: Used to filter pitching statistics based on the statistic inputted. 
          responses:
            "200":
              description: OK
              content:
                application/json:
                  schema:
                    type: array
    /relief_pitching_stats:
      get:
          operationId: getReliefPitchingStats
          summary: Retrieves a specific pitching statistic for all relief pitchers from Fangraphs for the season being inputted.
          parameters:
            - in: query
              name: year
              required: true
              schema: 
                type: integer
              description: Used to select which season the pitching statistics will be retrieved from.
            - in: query
              name: pitching_stat
              required: true
              schema: 
                type: string
                enum: ['W', 'L', 'WAR', 'ERA', 'GS', 'IP', 'HR', 'BB' , 'SO', 'GB', 'FB', 'LD', 'K/9', 'BB/9', 'H/9', 'AVG',
                      'WHIP', 'BABIP', 'LOB%', 'FIP']
              description: Used to filter pitching statistics based on the statistic inputted.
          responses:
            "200":
              description: OK
              content:
                application/json:
                  schema: 
                    type: array
    /pitching_stats_individual:
      get:
          operationId: getPitchingStatsIndividual
          summary: Retrieves the pitching stats for an individual player for a given season
          parameters:
            - in: query
              name: year
              required: true
              schema: 
                type: integer
              description: Used to select which season the pitching statisitics will be retrieved from
            - in: query
              name: key_fangraphs
              required: true
              schema: 
                type: integer
              description: >
                The fangraphs player identifier. This value can be retrieved by invoking 
                the operation with operationId 'getPlayerIdLookup'.
          responses:
            "200":
              description: OK
              content:
                application/json:
                  schema:
                    type: array
    /team_batting:
      get:
          operationId: getTeamBatting
          summary: >
            Retrieves the batting statistics from Baseball Reference for each player on the team for the provided season.
            This function should be used when you want to retrieve each individual player's batting statistics for a team. 
          parameters:
            - in: query
              name: team_abbreviation
              required: true
              schema:
                type: string
              description: The 3 letter abbreviated name of the baseball team.
              example: NYY
            - in: query
              name: year
              required: true
              schema: 
                type: integer
              description: Used to select which season the team batting statistics will be retrieved from.
          responses:
            "200":
              description: A list of player statistics for the queried team
              content:
                application/json:
                  schema: 
                    type: array
                    items:
                      $ref: '#/components/schemas/getTeamBattingResponse'
    /team_batting_combined:
        get:
            operationId: getTeamBattingCombined
            summary: >
              Retrieves the combined batting statistics for all teams across the MLB from Fangraphs for the specified season.
              This function should be used whenever a prompt is asking for combined statistics. 
            parameters:
              - in: query
                name: year
                required: true
                schema: 
                  type: integer
                description: The year from which the batting statistics should be retrieved from
              - in: query
                name: team_abbreviation
                required: false
                schema:
                  type: string
                description: >
                  The 3 letter abbreviated name of the baseball team. This parameter should be passed 
                  to get the combined statistics for a specific team. 
                example: NYY
              - in: query 
                name: batting_stat
                required: false
                schema: 
                  type: string
                  enum: ['H', '2B', '3B', 'HR', 'RBI', 'BB', 'IBB', 'SO', 'HBP', 'SH', 'SF', 'GDP', 'SB', 'CS', 'AVG', 
                      'OBP', 'SLG', 'OPS']
                description: >
                  Can be used to filter based on a certain batting statistic. 
                  Calling the function without this parameter can be used to get an overall view of stats returned from the function. 
            responses:
              "200":
                description: OK
                content:
                  application/json:
                    schema:
                      type: array
                      items:
                        $ref: '#/components/schemas/getTeamBattingCombinedResponse'
    /team_fielding:
      get: 
          operationId: getTeamFielding
          summary: Retrieves the fielding statistics from Baseball Reference for each player on the team for the provided season
          parameters:
            - in: query
              name: team_abbreviation
              required: true
              schema:
                type: string
              description: The 3 letter abbreviated name of the baseball team.
              example: NYY
            - in: query
              name: year
              required: true
              schema: 
                type: integer
              description: Used to select which season the team fielding statistics will be retrieved from. 
          responses: 
            "200":
              description: A complex JSON object response
              content:
                application/json:
                  schema: 
                    type: object
    /team_pitching:
      get: 
        operationId: getTeamPitching
        summary: Retrieves the pitching statistics from Baseball Reference for each player on the team for the provided season
        parameters:
          - in: query
            name: team_abbreviation
            required: true
            description: The 3 letter abbreviated name of the baseball team.
            schema:
              type: string
            example: NYY
          - in: query
            name: year
            required: true
            description: Used to select which season the team fielding statistics will be retrieved from
            schema: 
              type: integer
        responses:
          "200":
            description: A complex JSON object response
            content:
              application/json:
                schema: 
                  type: object
    /team_pitching_combined:
        get:
            operationId: getTeamPitchingCombined
            summary: >
              Retrieves the combined pitching statistics for all teams across the MLB from Fangraphs for the specified season.
              This function should be used whenever a prompt is asking for combined pitching statistics for a team. 
            parameters:
              - in: query
                name: year
                required: true
                schema: 
                  type: integer
                description: The year from which the pitching statistics should be retrieved from
              - in: query
                name: team_abbreviation
                required: false
                schema:
                  type: string
                description: >
                  The 3 letter abbreviated name of the baseball team. This parameter should be passed 
                  to get the combined statistics for a specific team. 
                example: NYY
            responses:
              "200":
                description: OK
                content:
                  application/json:
                    schema:
                      type: array
                      items:
                        type: object
    /playerid_lookup:
        get:
            operationId: getPlayerIdLookup
            summary: Retrieves a player ID value to be used with other data retrieval functions. 
            parameters:
              - in: query
                name: first
                required: true
                description: A player's first name.
                schema: 
                  type: string 
              - in: query
                name: last
                required: true
                description: A player's last name.
                schema: 
                  type: string
            responses:
              "200":
                description: >
                  A player IDs should be returned. This ID can be used to retrieve other statistics from functions. 
                  This ID will not be useful for a user. The model should use this for querying individual player statistics.
              "500":
                description: Error. Multiple players and their IDs are returned. Display the results in a markdown table. 
...
components:
  schemas:
...
    getTeamBattingResponse: 
      type: object
      properties:
        Pos:
          type: string
          description: The position abbreviation that the player plays
        Name:
          type: string
          description: The name of the player
        Year:
          type: number
          description: The year that is being queried
        Age:
          type: number
          description: The age of the player
        G:
          type: number
          description: The number of games the player has played in
        PA:
          type: number
          description: The number of plate appearances 
        AB:
          type: number
          description: The number of at-bats by the player
        R:
          type: number
          description: The player's number of runs for the season
        H:
          type: number
          description: The player's number of hits
        2B:
          type: number
          description: The player's number of doubles
        3B:
          type: number
          description: The player's number of triples
        HR:
          type: number
          description: The player's number of home runs
        RBI:
          type: number
          description: The number of runs batted in
        SB:
          type: number
          description: The number of stolen bases
        CS:
          type: number
          description: The number of times the player was caught stealing
        BB:
          type: number
          description: The number of walks
        SO:
          type: number
          description: 
        AVG:
          type: number
          description: The players batting average
        OBP:
          type: number
          description: 
        SLG:
          type: number
          description: 
        OPS:
          type: number
          description: 
        OPS+:
          type: number
          description: 
        TB:
          type: number
          description: 
        GDP:
          type: number
          description: 
        HBP:
          type: number
          description: 
        SH:
          type: number
          description: 
        SF:
          type: number
          description: 
        IBB:
          type: number
          description: 
...